Update runner tests
Use relative imports for runner components using the symbols from
datalad.runner

Split off cmd.py test code into a separate module.
